평균단가 특정 숫자를 계속 입력 시 어느 시점에서 0이 입력 되는 버그 수정

+++ app/src/main/kotlin/com/yjpapp/stockportfolio/util/StockUtils.kt
@@ -75,7 +75,7 @@ object StockUtils {
     fun getNumInsertComma(num: String): String {
         if (num.isEmpty()) return "0"
         val decimalFormat = DecimalFormat("###,###")
-        return decimalFormat.format(num.replace(",", "").toDouble())
+        return decimalFormat.format(num.replace(",", "").toBigDecimal())
     }
 
     //5000000 => $5,000,000 변환
